--- libs/model/src/services/commonProtocol/abi/contestAbi.ts
@@ -1,244 +1,0 @@
-export const contestABI = [
-  {
-    inputs: [
-      {
-        internalType: 'address',
-        name: 'creator',
-        type: 'address',
-      },
-      {
-        internalType: 'string',
-        name: 'url',
-        type: 'string',
-      },
-      {
-        internalType: 'bytes',
-        name: 'data',
-        type: 'bytes',
-      },
-    ],
-    name: 'addContent',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: '',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'uint256',
-        name: 'contentId',
-        type: 'uint256',
-        indexed: true,
-      },
-      {
-        internalType: 'address',
-        name: 'creator',
-        type: 'address',
-        indexed: true,
-      },
-      {
-        internalType: 'string',
-        name: 'url',
-        type: 'string',
-        indexed: false,
-      },
-    ],
-    type: 'event',
-    name: 'ContentAdded',
-    anonymous: false,
-  },
-  {
-    inputs: [
-      {
-        internalType: 'address',
-        name: 'voter',
-        type: 'address',
-      },
-      {
-        internalType: 'uint256',
-        name: 'id',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'nonpayable',
-    type: 'function',
-    name: 'voteContent',
-  },
-  {
-    inputs: [],
-    stateMutability: 'view',
-    type: 'function',
-    name: 'endTime',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: '',
-        type: 'uint256',
-      },
-    ],
-  },
-  {
-    inputs: [],
-    stateMutability: 'view',
-    type: 'function',
-    name: 'startTime',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: '',
-        type: 'uint256',
-      },
-    ],
-  },
-  {
-    inputs: [],
-    name: 'contestInterval',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: '',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'currentContentId',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: '',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'contestToken',
-    outputs: [
-      {
-        internalType: 'address',
-        name: '',
-        type: 'address',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'getWinnerIds',
-    outputs: [
-      {
-        internalType: 'uint256[]',
-        name: '',
-        type: 'uint256[]',
-      },
-    ],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'uint256',
-        name: 'contestId',
-        type: 'uint256',
-      },
-    ],
-    name: 'getPastWinners',
-    outputs: [
-      {
-        internalType: 'uint256[]',
-        name: '',
-        type: 'uint256[]',
-      },
-    ],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'uint256',
-        name: '',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-    name: 'content',
-    outputs: [
-      {
-        internalType: 'string',
-        name: 'url',
-        type: 'string',
-      },
-      {
-        internalType: 'uint256',
-        name: 'cumulativeVotes',
-        type: 'uint256',
-      },
-      {
-        internalType: 'address',
-        name: 'creator',
-        type: 'address',
-      },
-      {
-        internalType: 'bool',
-        name: 'completed',
-        type: 'bool',
-      },
-    ],
-  },
-  {
-    inputs: [],
-    name: 'FeeMangerAddress',
-    outputs: [
-      {
-        internalType: 'address',
-        name: '',
-        type: 'address',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'contestLength',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: '',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'newContest',
-    outputs: [],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    inputs: [],
-    name: 'endContest',
-    outputs: [],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-];
