+++ libs/model/src/services/commonProtocol/contestHelper.ts
@@ -3,8 +3,6 @@ import { commonProtocol } from '@hicommonwealth/shared';
 import { Mutex } from 'async-mutex';
 import Web3, { PayableCallOptions } from 'web3';
 import { AbiItem } from 'web3-utils';
-import { contestABI } from './abi/contestAbi';
-import { feeManagerABI } from './abi/feeManagerAbi';
 import { createWeb3Provider } from './utils';
 
 const nonceMutex = new Mutex();
@@ -52,7 +50,7 @@ const addContent = async (
     web3 = await createWeb3Provider(rpcNodeUrl);
   }
   const contestInstance = new web3.eth.Contract(
-    contestABI as AbiItem[],
+    commonProtocol.contestAbi as AbiItem[],
     contest,
   );
 
@@ -114,7 +112,7 @@ const voteContent = async (
     web3 = await createWeb3Provider(rpcNodeUrl);
   }
   const contestInstance = new web3.eth.Contract(
-    contestABI as AbiItem[],
+    commonProtocol.contestAbi as AbiItem[],
     contest,
   );
 
@@ -154,7 +152,7 @@ export const getContestStatus = async (
 ): Promise<ContestStatus> => {
   const web3 = new Web3(rpcNodeUrl);
   const contestInstance = new web3.eth.Contract(
-    contestABI as AbiItem[],
+    commonProtocol.contestAbi as AbiItem[],
     contest,
   );
 
@@ -191,7 +189,7 @@ export const getContestScore = async (
 ): Promise<ContestScores> => {
   const web3 = new Web3(rpcNodeUrl);
   const contestInstance = new web3.eth.Contract(
-    contestABI as AbiItem[],
+    commonProtocol.contestAbi as AbiItem[],
     contest,
   );
 
@@ -244,7 +242,7 @@ export const getContestBalance = async (
   const web3 = new Web3(rpcNodeUrl);
 
   const contestInstance = new web3.eth.Contract(
-    contestABI as AbiItem[],
+    commonProtocol.contestAbi as AbiItem[],
     contest,
   );
 
@@ -252,7 +250,7 @@ export const getContestBalance = async (
     contestInstance,
     contest,
     web3,
-    feeManagerABI,
+    commonProtocol.feeManagerAbi,
     oneOff,
   );
 
@@ -337,7 +335,7 @@ export const rollOverContest = async (
   return nonceMutex.runExclusive(async () => {
     const web3 = await createWeb3Provider(rpcNodeUrl);
     const contestInstance = new web3.eth.Contract(
-      contestABI as AbiItem[],
+      commonProtocol.contestAbi as AbiItem[],
       contest,
     );
