+++ libs/model/src/services/commonProtocol/contractHelpers.ts
@@ -8,7 +8,6 @@ import Web3 from 'web3';
 import { AbiItem } from 'web3-utils';
 
 import { Balances, TokenAttributes, getBalances } from '../tokenBalanceCache';
-import { contestABI } from './abi/contestAbi';
 
 /**
  * Retrieves a namespace.
@@ -97,7 +96,10 @@ export const getTokenAttributes = async (
   const web3 = new Web3(rpcNodeUrl);
   let addr = address;
   if (fetchFromContest) {
-    const contest = new web3.eth.Contract(contestABI as AbiItem[], address);
+    const contest = new web3.eth.Contract(
+      commonProtocol.contestAbi as AbiItem[],
+      address,
+    );
     addr = await contest.methods.contestToken().call();
   }
   if (addr === ZERO_ADDRESS) {
