+++ libs/shared/src/commonProtocol/abis/communityStakesAbi.ts
