+++ libs/shared/src/commonProtocol/abis/namespaceAbi.ts
@@ -1,4 +1,4 @@
-export const NamespaceAbi = [
+export const namespaceAbi = [
   {
     type: 'function',
     name: 'balanceOf',
