+++ libs/shared/src/commonProtocol/abis/tokenCommunityManagerAbi.ts
