+++ libs/shared/src/commonProtocol/index.ts
@@ -1,3 +1,12 @@
+export * from './abis/communityStakesAbi';
+export * from './abis/contestAbi';
+export * from './abis/erc20Abi';
+export * from './abis/feeManagerAbi';
+export * from './abis/lpBondingCurveAbi';
+export * from './abis/namespaceAbi';
+export * from './abis/namespaceFactoryAbi';
+export * from './abis/reservationHookAbi';
+export * from './abis/tokenCommunityManagerAbi';
 export * from './chainConfig';
 export * from './contractHelpers/Contest';
 export * from './contractHelpers/Launchpad';
