--- packages/commonwealth/client/scripts/helpers/ContractHelpers/Abi/ERC20Abi.ts
@@ -1,26 +1,0 @@
-export const Erc20Abi = [
-  {
-    inputs: [
-      { internalType: 'address', name: 'spender', type: 'address' },
-      { internalType: 'uint256', name: 'amount', type: 'uint256' },
-    ],
-    name: 'approve',
-    outputs: [{ internalType: 'bool', name: '', type: 'bool' }],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    inputs: [{ internalType: 'address', name: 'account', type: 'address' }],
-    stateMutability: 'view',
-    type: 'function',
-    name: 'balanceOf',
-    outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-  },
-  {
-    inputs: [],
-    stateMutability: 'view',
-    type: 'function',
-    name: 'decimals',
-    outputs: [{ internalType: 'uint8', name: '', type: 'uint8' }],
-  },
-];
