--- packages/commonwealth/client/scripts/helpers/ContractHelpers/Abi/LpBondingCurveAbi.ts
@@ -1,77 +1,0 @@
-export const LpBondingCurve = [
-  {
-    inputs: [
-      { internalType: 'address', name: 'tokenAddress', type: 'address' },
-      { internalType: 'uint256', name: 'minAmountOut', type: 'uint256' },
-    ],
-    stateMutability: 'payable',
-    type: 'function',
-    name: 'buyToken',
-    outputs: [{ internalType: 'bool', name: '', type: 'bool' }],
-  },
-  {
-    inputs: [
-      { internalType: 'address', name: 'tokenAddress', type: 'address' },
-      { internalType: 'uint256', name: 'amount', type: 'uint256' },
-      { internalType: 'uint256', name: 'minAmountOut', type: 'uint256' },
-    ],
-    stateMutability: 'payable',
-    type: 'function',
-    name: 'sellToken',
-  },
-  {
-    type: 'function',
-    name: 'getPrice',
-    inputs: [
-      { name: 'tokenAddress', type: 'address', internalType: 'address' },
-      { name: 'amountIn', type: 'uint256', internalType: 'uint256' },
-      { name: 'isBuy', type: 'bool', internalType: 'bool' },
-    ],
-    outputs: [{ name: '', type: 'uint256', internalType: 'uint256' }],
-    stateMutability: 'view',
-  },
-  {
-    inputs: [
-      { internalType: 'address', name: 'tokenAddress', type: 'address' },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-    name: '_getFloatingTokenSupply',
-    outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-  },
-  {
-    inputs: [{ internalType: 'address', name: '', type: 'address' }],
-    stateMutability: 'view',
-    type: 'function',
-    name: 'liquidity',
-    outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-  },
-  {
-    type: 'function',
-    name: '_launchpadLiquidity',
-    inputs: [
-      { name: 'tokenAddress', type: 'address', internalType: 'address' },
-    ],
-    outputs: [{ name: '', type: 'uint256', internalType: 'uint256' }],
-    stateMutability: 'view',
-  },
-  {
-    type: 'function',
-    name: 'transferLiquidity',
-    inputs: [
-      { name: 'tokenAddress', type: 'address', internalType: 'address' },
-      { name: 'minAmountOut', type: 'uint256', internalType: 'uint256' },
-    ],
-    outputs: [],
-    stateMutability: 'payable',
-  },
-  {
-    inputs: [
-      { internalType: 'address', name: 'tokenAddress', type: 'address' },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-    name: '_poolLiquidity',
-    outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-  },
-];
