--- packages/commonwealth/client/scripts/helpers/ContractHelpers/Abi/NamespaceFactoryAbi.ts
@@ -1,216 +1,0 @@
-export const namespaceFactoryAbi = [
-  {
-    type: 'function',
-    name: 'newSingleERC20Contest',
-    inputs: [
-      { name: 'name', type: 'string', internalType: 'string' },
-      { name: 'length', type: 'uint256', internalType: 'uint256' },
-      { name: 'winnerShares', type: 'uint256[]', internalType: 'uint256[]' },
-      { name: 'token', type: 'address', internalType: 'address' },
-      { name: 'voterShare', type: 'uint256', internalType: 'uint256' },
-      { name: 'exhangeToken', type: 'address', internalType: 'address' },
-    ],
-    outputs: [{ name: '', type: 'address', internalType: 'address' }],
-    stateMutability: 'nonpayable',
-  },
-  {
-    inputs: [],
-    stateMutability: 'view',
-    type: 'function',
-    name: 'reservationHook',
-    outputs: [
-      {
-        internalType: 'address',
-        name: '',
-        type: 'address',
-      },
-    ],
-  },
-  {
-    inputs: [
-      {
-        internalType: 'string',
-        name: 'name',
-        type: 'string',
-      },
-      {
-        internalType: 'string',
-        name: '_uri',
-        type: 'string',
-      },
-      {
-        internalType: 'address',
-        name: '_feeManager',
-        type: 'address',
-      },
-      {
-        internalType: 'bytes',
-        name: '_signature',
-        type: 'bytes',
-      },
-    ],
-    name: 'deployNamespace',
-    outputs: [],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'string',
-        name: 'name',
-        type: 'string',
-      },
-      {
-        internalType: 'string',
-        name: 'tokenName',
-        type: 'string',
-      },
-      {
-        internalType: 'uint256',
-        name: 'id',
-        type: 'uint256',
-      },
-      {
-        internalType: 'address',
-        name: 'exchangeToken',
-        type: 'address',
-      },
-      {
-        internalType: 'uint256',
-        name: 'scalar',
-        type: 'uint256',
-      },
-      {
-        internalType: 'uint256',
-        name: 'curve',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'nonpayable',
-    type: 'function',
-    name: 'configureCommunityStakeId',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'bytes32',
-        name: '',
-        type: 'bytes32',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-    name: 'getNamespace',
-    outputs: [
-      {
-        internalType: 'address',
-        name: '',
-        type: 'address',
-      },
-    ],
-  },
-  {
-    inputs: [
-      {
-        internalType: 'string',
-        name: 'name',
-        type: 'string',
-      },
-      {
-        internalType: 'uint256',
-        name: 'interval',
-        type: 'uint256',
-      },
-      {
-        internalType: 'uint256[]',
-        name: 'winnerShares',
-        type: 'uint256[]',
-      },
-      {
-        internalType: 'uint256',
-        name: 'id',
-        type: 'uint256',
-      },
-      {
-        internalType: 'uint256',
-        name: 'prizeShare',
-        type: 'uint256',
-      },
-      {
-        internalType: 'uint256',
-        name: 'voterShare',
-        type: 'uint256',
-      },
-      {
-        internalType: 'uint256',
-        name: 'feeShare',
-        type: 'uint256',
-      },
-      {
-        internalType: 'uint256',
-        name: 'weight',
-        type: 'uint256',
-      },
-    ],
-    name: 'newContest',
-    outputs: [
-      {
-        internalType: 'address',
-        name: '',
-        type: 'address',
-      },
-    ],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-  {
-    inputs: [
-      {
-        internalType: 'string',
-        name: 'name',
-        type: 'string',
-      },
-      {
-        internalType: 'uint256',
-        name: 'length',
-        type: 'uint256',
-      },
-      {
-        internalType: 'uint256[]',
-        name: 'winnerShares',
-        type: 'uint256[]',
-      },
-      {
-        internalType: 'uint256',
-        name: 'id',
-        type: 'uint256',
-      },
-      {
-        internalType: 'uint256',
-        name: 'voterShare',
-        type: 'uint256',
-      },
-      {
-        internalType: 'uint256',
-        name: 'weight',
-        type: 'uint256',
-      },
-      {
-        internalType: 'address',
-        name: 'exhangeToken',
-        type: 'address',
-      },
-    ],
-    name: 'newSingleContest',
-    outputs: [
-      {
-        internalType: 'address',
-        name: '',
-        type: 'address',
-      },
-    ],
-    stateMutability: 'nonpayable',
-    type: 'function',
-  },
-];
