--- packages/commonwealth/client/scripts/helpers/ContractHelpers/Abi/feeManagerAbi.ts
@@ -1,26 +1,0 @@
-export const feeManagerABI = [
-  {
-    inputs: [
-      {
-        internalType: 'address',
-        name: 'beneficiary',
-        type: 'address',
-      },
-      {
-        internalType: 'address',
-        name: 'token',
-        type: 'address',
-      },
-    ],
-    name: 'getBeneficiaryBalance',
-    outputs: [
-      {
-        internalType: 'uint256',
-        name: '',
-        type: 'uint256',
-      },
-    ],
-    stateMutability: 'view',
-    type: 'function',
-  },
-];
