+++ packages/commonwealth/client/scripts/helpers/ContractHelpers/Contest.ts
@@ -1,8 +1,5 @@
 import { ZERO_ADDRESS, commonProtocol } from '@hicommonwealth/shared';
 import { AbiItem, TransactionReceipt } from 'web3';
-import { ContestAbi } from './Abi/ContestAbi';
-import { Erc20Abi } from './Abi/ERC20Abi';
-import { feeManagerABI } from './Abi/feeManagerAbi';
 import ContractBase from './ContractBase';
 import NamespaceFactory from './NamespaceFactory';
 
@@ -13,7 +10,7 @@ class Contest extends ContractBase {
   namespaceFactory: NamespaceFactory;
 
   constructor(contractAddress: string, factoryAddress: string, rpc: string) {
-    super(contractAddress, ContestAbi, rpc);
+    super(contractAddress, commonProtocol.contestAbi, rpc);
     this.namespaceFactoryAddress = factoryAddress;
   }
 
@@ -202,7 +199,7 @@ class Contest extends ContractBase {
       }
     } else {
       const token = new this.web3.eth.Contract(
-        Erc20Abi as AbiItem[],
+        commonProtocol.erc20Abi as unknown as AbiItem[],
         tokenAddress,
       );
       const decimals = await token.methods.decimals().call();
@@ -244,7 +241,7 @@ class Contest extends ContractBase {
       this.contract,
       this.contractAddress,
       this.web3,
-      feeManagerABI,
+      commonProtocol.feeManagerAbi,
       oneOff,
     );
     return parseInt(contestBalance, 10);
